middleware/requirements.txt
# Calendar Sync Middleware - Python Dependencies
# Version: 1.0.0

# ============================================================================
# Core Web Framework
# ============================================================================
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# ============================================================================
# Database
# ============================================================================
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# ============================================================================
# CalDAV Client
# ============================================================================
caldav==1.3.9
icalendar==5.0.11
pytz==2024.1
python-dateutil==2.8.2

# ============================================================================
# HTTP and API
# ============================================================================
requests==2.31.0
urllib3==2.1.0

# ============================================================================
# Data Validation and Serialization
# ============================================================================
pydantic==2.5.3
marshmallow==3.20.1

# ============================================================================
# Environment and Configuration
# ============================================================================
python-dotenv==1.0.0

# ============================================================================
# Logging and Monitoring
# ============================================================================
python-json-logger==2.0.7
colorlog==6.8.0

# ============================================================================
# Utilities
# ============================================================================
click==8.1.7
python-slugify==8.0.1

# ============================================================================
# Security
# ============================================================================
cryptography==41.0.7

# ============================================================================
# Testing (optional, for development)
# ============================================================================
# pytest==7.4.3
# pytest-flask==1.3.0
# pytest-cov==4.1.0
# faker==21.0.0

# ============================================================================
# Code Quality (optional, for development)
# ============================================================================
# black==23.12.1
# flake8==7.0.0
# mypy==1.7.1
# pylint==3.0.3
